+++ src/NitroHttp.Cli/Views/Components/ResponseView.cs
@@ -24,7 +24,7 @@ public class ResponseView(
     /// <param name="responseStatus">The HTTP response status code.</param>
     /// <param name="responseCount">The number of returned items.</param>
     /// <param name="responseSize">The response size in bytes.</param>
-    public void Display(string requestUrl, string response, int responseStatus, int responseCount, int responseSize)
+    public void Display(string requestUrl, string response, int responseStatus, int responseCount, long responseSize)
     {
         var sw = Stopwatch.StartNew();
